Fixes the Encoder for Decimal, Date and DateTime
when using the `:drop_nil_fields` option, the `Encoder` was converting
all structs to maps including Decimal, Date and DateTime.

The new logic introduces the `Miss` library to convert the nested
structs to map and properly converting Decimal to float and Data and
DateTime to iso8601 string before continuing with the drop of nil
fields and encoding to Json.